Rename files to muon composite
