+++ tt-train/sources/ttml/CMakeLists.txt
@@ -74,8 +74,8 @@ set(SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/ops/sampling_op.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/adamw.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/adamw.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/muon.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/muon.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/muon_composite.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/muon_composite.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/optimizer_base.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/remote_optimizer.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/optimizers/sgd_composite.cpp
