+++ tt-train/tests/CMakeLists.txt
@@ -35,7 +35,7 @@ set(SOURCES
     optimizers/adamw_composite_test.cpp
     optimizers/sgd_test.cpp
     optimizers/adamw_test.cpp
-    optimizers/muon_test.cpp
+    optimizers/muon_composite_test.cpp
     ops/newton_schulz_op_test.cpp
     optimizers/adamw_full_precision_test.cpp
     optimizers/no_op_test.cpp
